// sanity/types.ts
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;
// ...
export type ProjectsQueryResult = Array<{
  _id: string;
  _updatedAt: string;
  name: string | null;
  completionDate: string | null;
  image: string | null;
  technologies: Array<{
    _id: string;
    _type: "technology";
    _createdAt: string;
    _updatedAt: string;
    _rev: string;
    name?: string;
    image?: {
      asset?: {
        _ref: string;
        _type: "reference";
        _weak?: boolean;
        [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
      };
      hotspot?: SanityImageHotspot;
      crop?: SanityImageCrop;
      alt?: string;
      _type: "image";
    };
    url?: string;
    imageUrl: string | null;
  }> | null;
  slug: string | null;
  projectUrl: string | null;
  repositoryUrl: string | null;
  elevatorPitch: string | null;
}>;